Ensure referring classes modified as a result of merges adhere to business rules
When a class that is the filler for a role restriction is merged, the referring
class may have it's role restrictions modified, and even deleted, .eg. in the case
that the surviving class is already a filler on another role. In these cases we must
ensure that the business rule governing the mininal number of N&S restrictions (2) still
holds.

Jira-1416