Add ability to cancel search
Restructured the use of the swing utility invoke later to prevent the GUI
from freezing while the search is running. Invoke later is now called after the
search returns and there's the need to update the results panel. The search button
switches to say cancel after it is pressed. At any time the user can cancel out
of the search.

Jira-1430